perf(ui/breadcrumb): optimize breadcrumb
