+++ packages/varlet-ui/src/breadcrumb/__tests__/index.spec.js
@@ -1,8 +1,8 @@
-import Breadcrumbs from '..'
+import breadcrumb from '..'
 import { createApp } from 'vue'
 import { mount } from '@vue/test-utils'
 
-test('test breadcrumbs plugin', () => {
-  const app = createApp({}).use(Breadcrumbs)
-  expect(app.component(Breadcrumbs.name)).toBeTruthy()
+test('test breadcrumb plugin', () => {
+  const app = createApp({}).use(breadcrumb)
+  expect(app.component(breadcrumb.name)).toBeTruthy()
 })
