--- packages/varlet-ui/src/breadcrumb/docs/zh-CN.md
@@ -1,112 +1,0 @@
-# 面包屑
-
-### 介绍
-
-一级一级的展示一些信息
-
-### 分隔符
-
-通过`separator`设置分隔符
-
-```html
-<template>
-  <var-space direction="column" size="large">
-    <var-breadcrumb>
-      <var-breadcrumbs>一级</var-breadcrumbs>
-      <var-breadcrumbs>二级</var-breadcrumbs>
-      <var-breadcrumbs>三级</var-breadcrumbs>
-    </var-breadcrumb>
-
-    <var-breadcrumb separator="\">
-      <var-breadcrumbs>一级</var-breadcrumbs>
-      <var-breadcrumbs>二级</var-breadcrumbs>
-      <var-breadcrumbs>三级</var-breadcrumbs>
-    </var-breadcrumb>
-  </var-space>
-</template>
-```
-
-### 每一级分隔符
-
-每一级的分隔符可以单独设置
-
-```html
-<template>
-  <var-breadcrumb separator="\">
-    <var-breadcrumbs>一级</var-breadcrumbs>
-    <var-breadcrumbs separator="~">二级</var-breadcrumbs>
-    <var-breadcrumbs>三级</var-breadcrumbs>
-</template>
-```
-
-### 分隔符插槽
-
-分隔符也支持插槽，可以深度定制
-
-```html
-<var-breadcrumb>
-  <var-breadcrumbs>
-    <template #separator>
-      <var-icon name="chevron-right" />
-    </template>
-    一级
-  </var-breadcrumbs>
-  <var-breadcrumbs>
-    <template #separator>
-      <var-icon name="chevron-right" />
-    </template>
-    二级
-  </var-breadcrumbs>
-  <var-breadcrumbs>
-    <template #separator>
-      <var-icon name="chevron-right" />
-    </template>
-    三级
-  </var-breadcrumbs>
-</var-breadcrumb>
-```
-
-## API
-
-### 属性
-
-#### Breadcrumb Props
-
-| 参数        | 说明   | 类型     | 默认值 |
-| ----------- | ------ | -------- | ------ |
-| `separator` | 分隔符 | _string_ | `/`    |
-
-#### breadcrumbs Props
-
-| 参数        | 说明   | 类型     | 默认值 |
-| ----------- | ------ | -------- | ------ |
-| `separator` | 分隔符 | _string_ | `/`    |
-
-### 插槽
-
-#### Breadcrumb Slots
-
-| 参数      | 说明       | 参数 |
-| --------- | ---------- | ---- |
-| `default` | 面包屑内容 | `-`  |
-
-#### breadcrumbs Slots
-
-| 参数      | 说明                     | 参数 |
-| --------- | ------------------------ | ---- |
-| `default` | 面包屑每一级的内容       | `-`  |
-| `default` | 面包屑每一级的分隔符内容 | `-`  |
-
-### 样式变量
-
-以下为组件使用的 css 变量，可以使用 [StyleProvider 组件](#/zh-CN/style-provider) 进行样式定制
-
-#### breadcrumbs Variables
-
-| 变量名                                   | 默认值                 |
-| ---------------------------------------- | ---------------------- |
-| `--breadcrumbs-active-color`         | `var(--color-primary)` |
-| `--breadcrumbs-active-hover-color`   | `#233dd2`              |
-| `--breadcrumbs-text-color`           | `#888`                 |
-| `--breadcrumbs-separator-margin`     | `0 6px`                |
-| `---breadcrumbs-separator-font-size` | `14px`                 |
