+++ packages/varlet-ui/src/breadcrumb/example/index.vue
@@ -1,56 +1,12 @@
 <script setup>
+import Varbreadcrumb from '../index'
 import { watchLang, AppType } from '@varlet/cli/client'
-import VarBreadcrumb from '..'
-import VarBreadcrumbs from '../../breadcrumbs'
-import VarSpace from '../../space'
-import VarIcon from '../../icon'
 import { use, pack } from './locale'
 
 watchLang(use)
 </script>
 
 <template>
-  <app-type>{{ pack.separator }}</app-type>
-  <var-space direction="column" size="large">
-    <var-breadcrumb>
-      <var-breadcrumbs>{{ pack.level1 }}</var-breadcrumbs>
-      <var-breadcrumbs>{{ pack.level2 }}</var-breadcrumbs>
-      <var-breadcrumbs>{{ pack.level3 }}</var-breadcrumbs>
-    </var-breadcrumb>
-
-    <var-breadcrumb separator="\">
-      <var-breadcrumbs>{{ pack.level1 }}</var-breadcrumbs>
-      <var-breadcrumbs>{{ pack.level2 }}</var-breadcrumbs>
-      <var-breadcrumbs>{{ pack.level3 }}</var-breadcrumbs>
-    </var-breadcrumb>
-  </var-space>
-
-  <app-type>{{ pack.levelSeparator }}</app-type>
-  <var-breadcrumb>
-    <var-breadcrumbs>{{ pack.level1 }}</var-breadcrumbs>
-    <var-breadcrumbs separator="~">{{ pack.level2 }}</var-breadcrumbs>
-    <var-breadcrumbs>{{ pack.level3 }}</var-breadcrumbs>
-  </var-breadcrumb>
-
-  <app-type>{{ pack.slotSeparator }}</app-type>
-  <var-breadcrumb>
-    <var-breadcrumbs>
-      <template #separator>
-        <var-icon name="chevron-right" />
-      </template>
-      {{ pack.level1 }}
-    </var-breadcrumbs>
-    <var-breadcrumbs>
-      <template #separator>
-        <var-icon name="chevron-right" />
-      </template>
-      {{ pack.level2 }}
-    </var-breadcrumbs>
-    <var-breadcrumbs>
-      <template #separator>
-        <var-icon name="chevron-right" />
-      </template>
-      {{ pack.level3 }}
-    </var-breadcrumbs>
-  </var-breadcrumb>
+  <app-type>Mobile phone example code</app-type>
+  <var-breadcrumb />
 </template>
