+++ packages/varlet-ui/src/breadcrumb/index.ts
@@ -1,11 +1,11 @@
 // Component entry, the folder where the file exists will be exposed to the user
+import breadcrumb from './Breadcrumb.vue'
 import type { App } from 'vue'
-import Breadcrumb from './Breadcrumb.vue'
 
-Breadcrumb.install = function (app: App) {
-  app.component(Breadcrumb.name, Breadcrumb)
+breadcrumb.install = function (app: App) {
+  app.component(breadcrumb.name, breadcrumb)
 }
 
-export const _BreadcrumbComponent = Breadcrumb
+export const _breadcrumbComponent = breadcrumb
 
-export default Breadcrumb
+export default breadcrumb
