+++ packages/varlet-ui/src/breadcrumb/props.ts
@@ -1,6 +1,5 @@
 export const props = {
   separator: {
     type: String,
-    default: '/',
   },
 }
