+++ packages/varlet-ui/src/breadcrumbs/__tests__/index.spec.js
@@ -1,8 +1,8 @@
-import breadcrumbs from '..'
+import Breadcrumbs from '..'
 import { createApp } from 'vue'
 import { mount } from '@vue/test-utils'
 
-test('test breadcrumbs plugin', () => {
-  const app = createApp({}).use(breadcrumbs)
-  expect(app.component(breadcrumbs.name)).toBeTruthy()
+test('test breadcrumb plugin', () => {
+  const app = createApp({}).use(Breadcrumbs)
+  expect(app.component(Breadcrumbs.name)).toBeTruthy()
 })
