+++ packages/varlet-ui/src/breadcrumbs/example/locale/zh-CN.ts
@@ -1,3 +1,8 @@
 export default {
-  // Example locale
+  separator: '分隔符',
+  level1: '一级',
+  level2: '二级',
+  level3: '三级',
+  levelSeparator: '每一级分隔符',
+  slotSeparator: '分隔符插槽',
 }
