+++ packages/varlet-ui/src/breadcrumbs/props.ts
@@ -1,5 +1,6 @@
 export const props = {
   separator: {
     type: String,
+    default: '/',
   },
 }
